Moved ecies specific implementation out of ecies and added key rotation
